==> scripts_analysis/0_calculate_basin_area.ipynb <==
# This code is used to calculate based area based on basin shapefile

import geopandas as gpd

invector = '/home/h294liu/project/proj/5_summaCalib/5_calib_test/BowAtBanff/model/shapefiles/catchment/bow_distributed_elevation_zone.shp'
dst_crs = "ESRI:102008"
area_field = 'area_sqm'

in_gdf = gpd.read_file(invector)        # read vector file
in_gdf_prj = in_gdf.to_crs(dst_crs)     # convert projection
in_gdf_prj[area_field] = in_gdf_prj.area        # calculate each polygon area
total_area = in_gdf_prj[area_field].sum()   # calculate total area

print('Catchment total area is %f (m^2).'%(total_area))

import numpy as np
domain_channel_length = np.sqrt(total_area)
print('Catchment channel length is %f (m).'%(domain_channel_length))

shape_default = 2.5
nGRU = float(51)
GRU_area = total_area/nGRU  # mean GRU area in square meter
GRU_channel_length = np.sqrt(GRU_area)  # mean GRU chennel length in meter

# assume lower and upper runoff velocity
v_priori, v_lower, v_upper = 1, 0.1, 10 # unit: m/s

# calculate routingGammaScale default, lower and upper value.
param_default  = (GRU_channel_length/v_priori)/shape_default
param_min = (GRU_channel_length/v_upper)/shape_default
param_max = (GRU_channel_length/v_lower)/shape_default

print(GRU_channel_length)
print(param_default, param_min, param_max)

GRU_area
