==> scripts_analysis/5_check_multiple_calib_jobs.ipynb <==
This code checks the stauts of a priori model run for each gauge - complete or fatal error. 

import os,shutil,glob,codecs
import pandas as pd
import numpy as np

root_dir = '/home/h294liu/project/proj/7_nelson'
gauge_csv = os.path.join(root_dir,'Flow_data/Phase_0&1_Natural/Natural_gauge_stn_info_calib_phase_0&1.csv')
gauge_df = pd.read_csv(gauge_csv)
output_basedir = '/home/h294liu/scratch/7_nelson/calib'

calib_type = 'DDS'
calib_num = 3 
incomplete_list = []

for index, row in gauge_df.iterrows():
    Gauge_ID = row['Gauge_ID']

    # Make multiple calib individually
    for calib_id in range(calib_num):
        
        calib_folder = Gauge_ID+'_'+calib_type+str(calib_id+1)
#         print(calib_folder)

        # work on gauge calib
        calib_dir = os.path.join(output_basedir,calib_folder)
        if not os.path.exists(calib_dir):
            os.makedirs(calib_dir)  

        out_files = glob.glob(os.path.join(calib_dir,'calib','*.out')) #'*.out'
        
        if out_files == []:
            print('%s does not have .out file yet.'%(calib_folder))
            continue
        else:
            if len(out_files) == 1:
                latest_out_file = out_files[0]  # when only one file for check
            else:
                latest_out_file = max(out_files, key=os.path.getctime) # when there are more than one files

            try:
                f = codecs.open(latest_out_file, encoding='utf-8', errors='strict')
                for line in f:
                    if 'OSError' in line:
                        incomplete_list.append(calib_folder)
                        print(Guage_ID)
                        print(line)
                        break
            except UnicodeDecodeError:
                incomplete_list.append(calib_folder)
                print(Guage_ID,"invalid utf-8")    
        
print('Incomplete model runs # = ',len(incomplete_list))
